refactored the loadzone view functions controller cleanup
